Texture System verbessert

+++ [GameEngine] TDP4 Team Battle Remake/src/me/xxfreakdevxx/de/game/Camera.java
@@ -3,9 +3,9 @@ package me.xxfreakdevxx.de.game;
 import me.xxfreakdevxx.de.game.object.GameObject;
 
 public class Camera {
-	private float x,y;
+	private double x,y;
 	
-	public Camera(float x, float y) {
+	public Camera(double x, double y) {
 		this.x=x;
 		this.y=y;
 	}
@@ -20,19 +20,19 @@ public class Camera {
 		if(x >= (Game.windowHeight + 16)) x = (Game.windowHeight+16);
 	}
 
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
